rafthttp: various clean up
